% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na2zero.R
\name{na2zero}
\alias{na2zero}
\title{FirstPackage}
\usage{
na2zero(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
Turns NAs in a dataframe into zeros
}
\examples{
dataset <- matrix(sample(c(NA, 1:5), 25, replace = TRUE), 5)
df <- as.data.frame(dataset)
FirstPackage::na2zero(df)

}
